lag komponent av antallBehandlinger
