+++ src/pages/saksbehandling/behandlingsoversikt/behandlingsoversikt-nb.ts
@@ -3,5 +3,4 @@ export default {
     ferdigeBehandlinger: 'Ferdige behandlinger',
     finnSak: 'Finn sak',
     nøkkeltall: 'Nøkkeltall',
-    behandlinger: 'behandlinger',
 };
