+++ src/pages/saksbehandling/behandlingsoversikt/ferdigeBehandlinger/FerdigeBehandlinger.tsx
@@ -1,12 +1,11 @@
 import * as RemoteData from '@devexperts/remote-data-ts';
-import { BodyShort, Label, Loader } from '@navikt/ds-react';
+import { Loader } from '@navikt/ds-react';
 import React, { useEffect, useState } from 'react';
 
 import ApiErrorAlert from '~components/apiErrorAlert/ApiErrorAlert';
 import * as sakSlice from '~features/saksoversikt/sak.slice';
 import { pipe } from '~lib/fp';
 import { useAsyncActionCreator } from '~lib/hooks';
-import { useI18n } from '~lib/i18n';
 import { Nullable } from '~lib/types';
 import {
     Filter,
@@ -18,12 +17,9 @@ import RestanserTabell from '~pages/saksbehandling/restans/Restanser';
 import { Restans, RestansStatus, RestansType } from '~types/Restans';
 import { toDateOrNull } from '~utils/date/dateUtils';
 
-import messages from '../behandlingsoversikt-nb';
-
-import styles from './ferdigeBehandlinger.module.less';
+import AntallBehandlinger from '../antallBehandlinger/AntallBehandlinger';
 
 export const FerdigeBehandlinger = () => {
-    const { formatMessage } = useI18n({ messages });
     const [hentFerdigeBehandlingerStatus, hentFerdigeBehandlinger] = useAsyncActionCreator(
         sakSlice.hentFerdigeBehandlinger
     );
@@ -62,14 +58,6 @@ export const FerdigeBehandlinger = () => {
             .filter((restans) =>
                 resultatfilter.length ? resultatfilter.includes(restans.status as keyof RestansResultatFilter) : true
             );
-    };
-
-    const AntallBehandlinger = (args: { restanser: Restans[] }) => {
-        return args.restanser.length > 0 ? (
-            <div className={styles.antallBehandlingerContainer}>
-                <Label>{args.restanser.length}</Label> <BodyShort>{formatMessage('behandlinger')}</BodyShort>
-            </div>
-        ) : null;
     };
 
     return (
