--- src/pages/saksbehandling/behandlingsoversikt/åpneBehandlinger/åpneBehandlinger.module.less
@@ -1,6 +1,0 @@
-@import '~/styles/variables.less';
-
-.antallBehandlingerContainer {
-    display: flex;
-    gap: @spacing-xxs;
-}
